Stylecop updates
