+++ src/Altinn-application-owner-system/Functions/EventsConfirmation.cs
@@ -7,15 +7,25 @@ using Microsoft.Extensions.Logging;
 
 namespace AltinnApplicationOwnerSystem.Functions
 {
-    public  class EventsConfirmation
+    /// <summary>
+    /// Azure Function that confirmes that data for a given instance is downloaded
+    /// </summary>
+    public class EventsConfirmation
     {
         private readonly IAltinnApp _altinnApp;
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="EventsConfirmation"/> class.
+        /// </summary>
         public EventsConfirmation(IAltinnApp altinnApp)
         {
             _altinnApp = altinnApp;
         }
 
+        /// <summary>
+        /// Function method that is triggered
+        /// </summary>
+        /// <returns>A <see cref="Task"/> representing the result of the asynchronous operation.</returns>
         [Function("EventsConfirmation")]
         public async Task Run([QueueTrigger("events-confirmation", Connection = "QueueStorage")]string item, FunctionContext executionContext)
         {
