+++ src/Altinn-application-owner-system/Functions/Services/Interface/IAuthenticationService.cs
@@ -5,9 +5,11 @@ using System.Threading.Tasks;
 
 namespace AltinnApplicationOwnerSystem.Functions.Services.Interface
 {
+    /// <summary>
+    /// Interface defining service responsible for the authentication process for Application Owner system
+    /// </summary>
     public interface IAuthenticationService
     {
-
         /// <summary>
         /// Methods that return Altinn token. If not cached it will log the solution
         /// in to MaskinPorten and then exchange the Maskinporten token to an Altinn token.
