+++ src/Altinn-application-owner-system/Functions/Services/Interface/IMaskinPortenClientWrapper.cs
@@ -6,8 +6,14 @@ using System.Threading.Tasks;
 
 namespace AltinnApplicationOwnerSystem.Functions.Services.Interface
 {
+    /// <summary>
+    /// Interface defining the MaskinPorten Client wrapper responsible for authentication of Application Owner System
+    /// </summary>
     public interface IMaskinPortenClientWrapper
     {
+        /// <summary>
+        /// Post MaskinPorten Authentication request
+        /// </summary>
         Task<string> PostToken(FormUrlEncodedContent bearer); 
     }
 }
