+++ src/Altinn-application-owner-system/Functions/Services/Interface/IQueueService.cs
@@ -1,5 +1,5 @@
-using AltinnApplicationOwnerSystem.Functions.Models;
 using System.Threading.Tasks;
+using AltinnApplicationOwnerSystem.Functions.Models;
 
 namespace AltinnApplicationOwnerSystem.Functions.Services.Interface
 {
@@ -14,7 +14,6 @@ namespace AltinnApplicationOwnerSystem.Functions.Services.Interface
         /// <param name="content">The content to push to the queue in string format</param>
         /// <returns>Returns a queue receipt</returns>
         public Task<PushQueueReceipt> PushToInboundQueue(string content);
-
 
         /// <summary>
         /// Pushes the provided content to the queue
